Merge feat-finished-35-related-Section-... into dev-arheo

+++ src/main/java/ro/ubb/abc2024/arheo/controller/section/SectionController.java
@@ -2,16 +2,31 @@ package ro.ubb.abc2024.arheo.controller.section;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import ro.ubb.abc2024.arheo.domain.artifact.Artifact;
+import ro.ubb.abc2024.arheo.domain.section.Section;
 import ro.ubb.abc2024.arheo.service.SectionService;
+import ro.ubb.abc2024.arheo.utils.converter.ArtifactDtoConverter;
 import ro.ubb.abc2024.arheo.utils.converter.SectionDtoConverter;
+import ro.ubb.abc2024.arheo.utils.dto.ArtifactDto;
 import ro.ubb.abc2024.arheo.utils.dto.SectionDto;
+import ro.ubb.abc2024.controller.UserController;
 import ro.ubb.abc2024.user.Role;
+import ro.ubb.abc2024.user.UserService;
+import ro.ubb.abc2024.user.UserServiceImpl;
 import ro.ubb.abc2024.utils.dto.Result;
+import ro.ubb.abc2024.arheo.controller.utils.HelperMethods;
+
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
@@ -20,42 +35,62 @@ import java.util.stream.Collectors;
 public class SectionController {
     private final SectionService sectionService;
     private final SectionDtoConverter sectionDtoConverter;
+    private final ArtifactDtoConverter artifactDtoConverter;
+    private final UserService userService;
 
     @GetMapping
-    //@PreAuthorize("hasAnyAuthority('ARH')")
-    // get all sections
-    public Result<List<SectionDto>> getSections() {
-        return new Result<>(true, HttpStatus.OK.value(), "Retrieved all sections", sectionService.getSections().stream()
-                .map(sectionDtoConverter::createFromEntity)
-                .collect(Collectors.toList()));
+    @PreAuthorize("hasAnyAuthority('SCOPE_ARH', 'SCOPE_ADMIN')")
+    public Result<Map<String, Object>> getSections(
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "5") int pageSize,
+            @RequestParam(required = false) Long sectionId,
+            @RequestParam(required = false) String sectionName,
+            @RequestParam(required = false) Long siteId,
+            @RequestParam(required = false) String status) {
+
+        Pageable pageable = PageRequest.of(page, pageSize);
+        Page<Section> sectionsPage = sectionService.findAllByCriteria(sectionId, sectionName, siteId, status, pageable);
+        Map<String, Object> response = HelperMethods.makeResponse(sectionsPage, sectionDtoConverter);
+        return new Result<>(true, HttpStatus.OK.value(), "Retrieved all sections based on params", response);
     }
 
     // get by id
     @GetMapping("/{id}")
-    //@PreAuthorize("hasAnyAuthority('ARH')")
+    @PreAuthorize("hasAnyAuthority('SCOPE_ARH', 'SCOPE_ADMIN')")
     public Result<SectionDto> getSection(@PathVariable long id) {
         return new Result<>(true, HttpStatus.OK.value(), "Retrieved section", sectionDtoConverter.createFromEntity(sectionService.getSection(id)));
     }
 
     @PostMapping
-    //@PreAuthorize("hasAnyAuthority('ARH')")
+    @PreAuthorize("hasAnyAuthority('SCOPE_ARH', 'SCOPE_ADMIN')")
     // add a new section, add it to the list of sections
     public Result<SectionDto> addSection(@RequestBody SectionDto sectionDto) {
+        var userId = userService.getUser(SecurityContextHolder.getContext().getAuthentication().getName()).getId();
+        var mainArchaeologistId = sectionService.getSection(sectionDto.siteId()).getSite().getMainArchaeologist().getId();
+        if (!Objects.equals(userId, mainArchaeologistId)) {
+            return new Result<>(false, HttpStatus.FORBIDDEN.value(), "Only the main archaeologist of the site can add sections", null);
+        }
         var section = sectionService.addSection(sectionDtoConverter.createFromDto(sectionDto));
-        //return sectionDtoConverter.createFromEntity(section);
         return new Result<>(true, HttpStatus.CREATED.value(), "Added section", sectionDtoConverter.createFromEntity(section));
     }
 
     @DeleteMapping
-    //@PreAuthorize("hasAnyAuthority('ARH')")
+    @PreAuthorize("hasAnyAuthority('SCOPE_ARH', 'SCOPE_ADMIN')")
     public Result<SectionDto> removeSection(@RequestBody long sectionId){
         // given a section id, remove the section from the list of sections
+        var section = sectionService.getSection(sectionId);
+        var userId = userService.getUser(SecurityContextHolder.getContext().getAuthentication().getName()).getId();
+        var mainArchaeologistId = sectionService.getSection(section.getSite().getMainArchaeologist().getId())
+                .getSite().getMainArchaeologist().getId();
+        if (!Objects.equals(userId, mainArchaeologistId)) {
+            return new Result<>(false, HttpStatus.FORBIDDEN.value(), "Only the main archaeologist of the site can add sections", null);
+        }
         sectionService.deleteSection(sectionId);
         return new Result<>(true, HttpStatus.OK.value(), "Deleted section", sectionDtoConverter.createFromEntity(sectionService.getSection(sectionId)));
     }
 
     @PutMapping
-    //@PreAuthorize("hasAnyAuthority('ARH')")
+    @PreAuthorize("hasAnyAuthority('SCOPE_ARH', 'SCOPE_ADMIN')")
     // edit/update a section; subject to change depending on what fields could be updated
     // could be improved by adding to SectionService a method that takes the Dto as parameter
     public Result<SectionDto> updateSection(@RequestBody SectionDto sectionDto) {
@@ -64,7 +99,7 @@ public class SectionController {
     }
 
     @GetMapping("/incomplete")
-    //@PreAuthorize("hasAnyAuthority('ARH')")
+    @PreAuthorize("hasAnyAuthority('SCOPE_ARH', 'SCOPE_ADMIN')")
     // get all sections that are incomplete
     public Result<List<SectionDto>> getIncompleteSections() {
         return new Result<>(true, HttpStatus.OK.value(), "Retrieved all incomplete sections", sectionService.getIncompleteSections().stream()
@@ -73,23 +108,21 @@ public class SectionController {
     }
 
     @GetMapping("/artefacts/{sectionId}")
-    //@PreAuthorize("hasAnyAuthority('ARH')")
+    @PreAuthorize("hasAnyAuthority('SCOPE_ARH', 'SCOPE_ADMIN')")
     // get all artifacts from a section
-    public Result<List<Long>> getArtifactsFromSection(@PathVariable long sectionId) {
-        return new Result<>(true, HttpStatus.OK.value(), "Retrieved all artifacts from section", sectionService.getArtifactsFromSection(sectionId)
-                .stream()
-                .map(Artifact::getId)
-                .collect(Collectors.toList())
-        );  //todo: return ArtifactDto once it is implemented (or not, depending on what we want)
+    public Result<List<ArtifactDto>> getArtifactsFromSection(@PathVariable long sectionId) {
+        return new Result<>(true, HttpStatus.OK.value(), "Retrieved all artifacts from section", sectionService.getArtifactsFromSection(sectionId).stream()
+                .map(artifactDtoConverter::createFromEntity)
+                .collect(Collectors.toList()));
     }
 
     @GetMapping("/artefacts")
-    //@PreAuthorize("hasAnyAuthority('ARH')")
+    @PreAuthorize("hasAnyAuthority('SCOPE_ARH', 'SCOPE_ADMIN')")
     // get all artifacts from a section, from a specific archaeologist
-    public Result<List<Long>> getArtifactsFromSectionByArchaeologist(@RequestParam long sectionId, @RequestParam long archaeologistId) {
-        return new Result<>(true, HttpStatus.OK.value(), "Retrieved all artifacts from section by archaeologist", sectionService.getArtifactsFromSectionByArchaeologist(sectionId, archaeologistId)
-                .stream()
-                .map(Artifact::getId)
+    public Result<List<ArtifactDto>> getArtifactsFromSectionByArchaeologist(@RequestParam long sectionId, @RequestParam long archaeologistId) {
+        return new Result<>(true, HttpStatus.OK.value(), "Retrieved all artifacts from section", sectionService.getArtifactsFromSectionByArchaeologist(sectionId, archaeologistId).
+                stream()
+                .map(artifactDtoConverter::createFromEntity)
                 .collect(Collectors.toList()));
     }
 }
